test(handlers): split projects.settings.test.ts along describe seams
Split the 679-line projects.settings.test.ts into four sibling files
along its existing top-level describe seams, each importing shared
fixtures from ./projects.test-helpers.ts (no copy-pasted setup):

- projects.warren-config.test.ts (warren-435b)
- projects.triggers.test.ts (warren-99c3)
- projects.triggers-run.test.ts (warren-99c3)
- projects.seed-status.test.ts (warren-4015)

All 17 tests preserved and green; each new file is well under the
500-line limit. Removes the now-stale grandfather entry from
scripts/file-size-budgets.json.

Closes warren-a715, warren-6d84
