+++ scripts/file-size-budgets.json
@@ -22,7 +22,6 @@
 		"src/server/bridges.test.ts": 557,
 		"src/server/handlers/agents.test.ts": 533,
 		"src/server/main/index.ts": 548,
-		"src/server/handlers/projects.settings.test.ts": 685,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
 		"src/server/integration.multi-worker.test.ts": 553,
 		"src/server/server.test.ts": 536
